Add re-export stubs for ProtectedRoute
